authorizer_service/requirements.txt
boto3==1.34.0
PyJWT==2.8.0
python-dateutil==2.8.2

